// src/app/models/antiPatternMasterList.ts
export class antiPatternMasterList {
  id: number;
  language: string;
  group: string;  
  subGroup: string;  
  desc: string;
  effortOneTime: number;
  effortPerMatch: number;
  effortPerFileSize: number;  
  severity: number;  
  migrationRecommendation4R: string;  
  recommendation: string;  
  enabled: boolean;  
}